% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{graf_pvo}
\alias{graf_pvo}
\title{Gráfico de Valores Preditos vs Observados}
\usage{
graf_pvo(modelo, x_var, fixas = list(), dados)
}
\arguments{
\item{modelo}{Saída de uma função de regressão linear}

\item{x_var}{Nome da variável correspondente ao eixo X}

\item{fixas}{Lista das outras variáveis do modelo que assumem um valor fixo especificado}

\item{dados}{Banco de dados utilizado na regressão linear feita previamente}
}
\value{
Retorna um gráfico de valores preditos vs valores observados, adicionado de uma reta de regressão calculada entre os dois vetores.
}
\description{
A função faz um gráfico cujos eixos correspondem aos valores observados da variável resposta de um banco de dados, e aos valores preditos através de uma regressão linear do mesma variável.
}
\details{
É necessário que x_var seja no formato Character e fixas seja uma lista na forma list(xi = valor, xj = valor).
}
\examples{
graf_pvo(modelo, x_var = "x1", fixas = list(x2 = 0.5, x3 = 0.5), dados = df)
graf_pvo(modelo, "x2", list(x1 = 0.7, x3 = 0.1), df)
graf_pvo(saida_regressao_linear, "x3", fixas = list(x1 = 0.5, x2 = 0.9), df)

}
