% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicts}
\alias{predicts}
\title{Predição de uma Regressão Linear}
\usage{
predicts(modelo, valores)
}
\arguments{
\item{modelo}{Saída de uma função de regressão linear}

\item{valores}{Matriz em que cada linha correponde aos valores preditos desejados, e cada coluna a variável preditora na ordem em que o resultado é dado na saída do modelo.}
}
\value{
Retorna uma lista com os resultados dos valores preditos
}
\description{
A função tem como objetivo fazer uma predição a partir de um modelo de regressão linear feito através de uma função de regressão linear
}
\details{
É necessário que os valores sejam numéricos e possuam pelo menos um valor de predição para cada variável pertencente ao modelo inserido na função.
}
\examples{
modelo = reg_lin(pacote2::df, "y", c("x1", "x2"))
valores = matrix(c(1, 5,
                   2, 6,
                   3, 7), nrow = 3, byrow = TRUE)
predicts(modelo, valores)

}
