% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglin.R
\name{reg_lin}
\alias{reg_lin}
\title{Cálculo de Regressão Linear}
\usage{
reg_lin(dados, Y, Xs)
}
\arguments{
\item{dados}{Banco de dados a ser utilizado}

\item{Y}{string com o nome da Variável resposta da forma como está no banco}

\item{Xs}{vetor de strings com os nomes das variáveis preditoras}
}
\value{
Retorna uma lista com 3 objetos: um dataframe de nome "coeficientes" com os valores dos betas calculados, um vetor "preditos" com os valores preditos pelo modelo e um vetor "residuos" com os erros dos valores preditos.
}
\description{
A função tem como objetivo aplicar o cálculo da regressão linear simples a um banco de dados
}
\details{
É necessário que tanto a variável resposta quanto as variáveis preditoras sejam numéricas e tenham o mesmo tamanho
}
\examples{
reg_lin(df, "y", c("x1", "x2", "x3"))
reg_lin(df, "y", c("x1", "x3"))
reg_lin(df, "y", c("x2"))

}
