% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicts}
\alias{predicts}
\title{Predição de uma Regressão Linear}
\usage{
predicts(modelo, valores)
}
\arguments{
\item{modelo}{saída da funçaõ reg_lin}

\item{valores}{matriz com uma linha para cada valor predito desejado e cada coluna o valor da variável preditora na ordem e que o resultado é dado na saída do modelo}
}
\value{
Retorna uma lista com os resultados dos valores preditos
}
\description{
A função tem como objetivo fazer uma prediçaõ a aprtir de um modelo de regressão linear feito através da função reg_lin
}
\details{
É necessário que
}
\examples{
modelo = reg_lin(pacote2::df, "y", c("x1", "x2"))
valores = matrix(c(1, 5,
                   2, 6,
                   3, 7), nrow = 3, byrow = TRUE)
predicts(modelo, valores)

}
